% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb.R
\name{pb}
\alias{pb}
\alias{pb.default}
\alias{pb.sp_iterable}
\title{Wrap an interable in a progress bar}
\usage{
pb(object, ...)

\method{pb}{default}(object, ...)

\method{pb}{sp_iterable}(object, ...)
}
\description{
Wrap an interable in a progress bar
}
